clippy fixes:logical plan mod
