+++ brahmand/src/query_planner/analyzer/query_validation.rs
@@ -8,7 +8,7 @@ use crate::{
             errors::{AnalyzerError, Pass},
         },
         logical_expr::logical_expr::Direction,
-        logical_plan::logical_plan::LogicalPlan,
+        logical_plan::LogicalPlan,
         plan_ctx::PlanCtx,
         transformed::Transformed,
     },
