+++ brahmand/src/query_planner/logical_plan/skip_n_limit_clause.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use crate::{
     open_cypher_parser::ast::{LimitClause, SkipClause},
-    query_planner::logical_plan::logical_plan::{Limit, LogicalPlan, Skip},
+    query_planner::logical_plan::{Limit, LogicalPlan, Skip},
 };
 
 pub fn evaluate_skip_clause(skip_clause: &SkipClause, plan: Arc<LogicalPlan>) -> Arc<LogicalPlan> {
