+++ brahmand/src/query_planner/logical_plan/where_clause.rs
@@ -4,7 +4,7 @@ use crate::{
     open_cypher_parser::ast::WhereClause,
     query_planner::{
         logical_expr::logical_expr::LogicalExpr,
-        logical_plan::logical_plan::{Filter, LogicalPlan},
+        logical_plan::{Filter, LogicalPlan},
     },
 };
