+++ brahmand/src/query_planner/optimizer/projection_push_down.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use crate::query_planner::{
-    logical_plan::logical_plan::{LogicalPlan, Projection},
+    logical_plan::{LogicalPlan, Projection},
     optimizer::{
         errors::{OptimizerError, Pass},
         optimizer_pass::{OptimizerPass, OptimizerResult},
